Fixed Alert, CheckboxGroup, and Information component imports to use SyllabusComponents versions in TestPageRunner.jsx.
