+++ frontend/package.json
@@ -30,10 +30,7 @@
     "build": "react-scripts build",
     "test": "cross-env CI=false react-scripts test",
     "eject": "react-scripts eject",
-    "myComponents": "node test-runner.js myComponents",
-    "checkboxes": "node test-runner.js checkboxes",
-    "alerts": "node test-runner.js alerts",
-    "all": "node test-runner.js all"
+    "myComponents": "node test-runner.js components"
   },
   "eslintConfig": {
     "extends": [
