+++ frontend/src/App.tsx
@@ -1,16 +1,17 @@
 import React from 'react';
+import { BrowserRouter } from 'react-router-dom';
 import AppRoutes from './navigation/AppRoutes'
 import {AuthProvider} from "./context/AuthContext";
 
 const App: React.FC = () => {
     return (
-        <>
+        <BrowserRouter>
             <main>
                 <AuthProvider>
                     <AppRoutes/>
                 </AuthProvider>
             </main>
-        </>
+        </BrowserRouter>
     );
 };
