+++ frontend/src/components/SyllabusComponents/CheckboxGroup.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 
 function CheckboxGroup({
-  id = "meeting_days_checkboxes",
+  id = "checkbox_group",
   data = ["M", "T", "W", "R", "F"],
-  horizontal = false,
+  horizontal = true,
 }) {
   // Exclude 'Check All' from initial checked state
   const items = data.filter((d) => d !== "Check All");
@@ -30,27 +30,41 @@ function CheckboxGroup({
     <div
       id={id}
       role="group"
-      aria-label="Meeting days"
+      aria-label="Checkbox group"
       style={{
         display: "flex",
         flexDirection: horizontal ? "row" : "column",
         gap: "0.5rem",
+        flexWrap: horizontal ? "wrap" : "nowrap",
       }}
     >
       {data.map((d) => (
         <label
           key={d}
-          style={{ display: "flex", alignItems: "center", gap: "0.25rem" }}
+          style={{
+            display: "flex",
+            alignItems: "center",
+            gap: "0.5rem",
+            cursor: "pointer",
+            padding: "0.25rem",
+            fontSize: "1rem",
+            fontWeight: d === "Check All" ? "600" : "400",
+            color: d === "Check All" ? "#007acc" : "#333",
+          }}
         >
           <input
             type="checkbox"
             value={d}
             checked={
               d === "Check All"
-                ? checked.length === items.length && items.length > 0
+                ? checked.length === items.length
                 : checked.includes(d)
             }
             onChange={() => handleChange(d)}
+            style={{
+              transform: "scale(1.2)",
+              marginRight: "0.25rem",
+            }}
           />
           {d}
         </label>
