+++ frontend/src/components/SyllabusComponents/Information.jsx
@@ -1,61 +1,54 @@
-// import React from "react";
+import React from "react";
+import { FaExclamationCircle } from "react-icons/fa";
 
-// // Font Awesome CDN link should be added in public/index.html or index.html for icons to work
-// <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css" />
+function Information({ text }) {
+  return (
+    <div
+      style={{
+        display: "flex",
+        alignItems: "center",
+        background: "#d4edda",
+        border: "2px solid #28a745",
+        borderRadius: "50px",
+        padding: "12px 24px 12px 16px",
+        margin: "1rem 0",
+        fontSize: "16px",
+        fontWeight: "400",
+        color: "#155724",
+        lineHeight: "1.4",
+        maxWidth: "100%",
+        boxSizing: "border-box",
+      }}
+    >
+      {/* Exclamation Icon */}
+      <div
+        style={{
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "center",
+          width: "32px",
+          height: "32px",
+          backgroundColor: "#fff",
+          borderRadius: "50%",
+          border: "2px solid #28a745",
+          marginRight: "12px",
+          flexShrink: 0,
+        }}
+      >
+        <FaExclamationCircle 
+          style={{ 
+            color: "#dc3545", 
+            fontSize: "18px" 
+          }}
+        />
+      </div>
+      
+      {/* Text content */}
+      <span style={{ flex: 1 }}>
+        {text}
+      </span>
+    </div>
+  );
+}
 
-// function Information({ text }) {
-//   return (
-//     <div
-//       style={{
-//         display: "flex",
-//         alignItems: "center",
-//         background: "#d2f4e1",
-//         border: "3px solid #2ecc71",
-//         borderRadius: "2rem",
-//         padding: "0.5rem 1.5rem 0.5rem 1rem",
-//         color: "#222",
-//         fontSize: "1.25rem",
-//         fontWeight: 500,
-//         gap: "1rem",
-//         margin: "1rem 0"
-//       }}
-//     >
-//       <span style={{ color: "#c30052", fontSize: "2rem", display: "flex", alignItems: "center" }}>
-//         {/* <FontAwesomeIcon icon={byPrefixAndName.far['circle-exclamation']} /> */}
-//         <i class="fa-regular fa-circle-exclamation"></i>
-//         {/* <i class="fa-regular fa-circle-exclamation" aria-hidden="true"></i> */}
-//         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css" />
-//       return (
-//         <div
-//           style={{
-//             display: "flex",
-//             alignItems: "center",
-//             background: "#d2f4e1",
-//             border: "4px solid #2ecc71",
-//             borderRadius: "999px",
-//             padding: "0.5rem 1.5rem 0.5rem 1rem",
-//             color: "#222",
-//             fontSize: "1.25rem",
-//             fontWeight: 500,
-//             gap: "1rem",
-//             margin: "1rem 0"
-//           }}
-//         >
-//           <span style={{
-//             color: "#c30052",
-//             fontSize: "2rem",
-//             display: "flex",
-//             alignItems: "center",
-//             justifyContent: "center",
-//             width: "2.5rem",
-//             height: "2.5rem",
-//             background: "#fff",
-//             borderRadius: "50%",
-//             border: "4px solid #2ecc71",
-//             marginRight: "0.75rem"
-//           }}>
-//             <i className="fa-solid fa-circle-exclamation" aria-hidden="true"></i>
-//           </span>
-//           <span>{text}</span>
-//         </div>
-//       );
+export default Information;
