--- frontend/src/components/SyllabusComponents/Information_fixed.jsx
@@ -1,28 +1,0 @@
-import React from "react";
-
-function Information({ text }) {
-  return (
-    <div
-      style={{
-        display: "flex",
-        alignItems: "center",
-        background: "#d2f4e1",
-        border: "3px solid #2ecc71",
-        borderRadius: "2rem",
-        padding: "0.5rem 1.5rem 0.5rem 1rem",
-        color: "#222",
-        fontSize: "1.25rem",
-        fontWeight: 500,
-        gap: "1rem",
-        margin: "1rem 0"
-      }}
-    >
-      <span style={{ color: "#c30052", fontSize: "2rem", display: "flex", alignItems: "center" }}>
-        <i className="fa-solid fa-circle-exclamation" aria-hidden="true"></i>
-      </span>
-      <span>{text}</span>
-    </div>
-  );
-}
-
-export default Information;
