+++ frontend/src/index.js.backup
@@ -6,109 +6,16 @@ import '@fontsource/inter/700.css'
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
+import App from './App';
 import reportWebVitals from './reportWebVitals';
-
-// Simple, testable checkbox group component
-function CheckboxGroup({
-  id = "meeting_days_checkboxes",
-  data = ["M", "T", "W", "R", "F"],
-  horizontal = true,
-}) {
-  return (
-    <div
-      id={id}
-      role="group"
-      aria-label="Meeting days"
-      style={{
-        display: 'flex',
-        flexDirection: horizontal ? 'row' : 'column',
-        gap: '0.5rem',
-        margin: '1rem 0',
-      }}
-    >
-      {data.map((d) => (
-        <label
-          key={d}
-          style={{ display: 'flex', alignItems: 'center', gap: '0.25rem' }}
-        >
-          <input type="checkbox" value={d} />
-          {d}
-        </label>
-      ))}
-    </div>
-  );
-}
-
-// Component imports with error handling
-let CheckboxGroupTest, InformationTest, AlertB, InformationB;
-
-try {
-  CheckboxGroupTest = require('./components/screens/CheckboxGroupTest').default;
-} catch (e) {
-  console.warn('CheckboxGroupTest not found:', e.message);
-  CheckboxGroupTest = () => <div>CheckboxGroupTest component not found</div>;
-}
-
-try {
-  InformationTest = require('./components/screens/InformationTest').default;
-} catch (e) {
-  console.warn('InformationTest not found:', e.message);
-  InformationTest = () => <div>InformationTest component not found</div>;
-}
-
-try {
-  AlertB = require('./components/SyllabusComponents/AlertB').default;
-} catch (e) {
-  console.warn('AlertB not found:', e.message);
-  AlertB = ({ text }) => <div style={{ padding: '1rem', background: '#f0f0f0', border: '1px solid #ccc' }}>AlertB: {text}</div>;
-}
-
-try {
-  InformationB = require('./components/SyllabusComponents/InformationB').default;
-} catch (e) {
-  console.warn('InformationB not found:', e.message);
-  InformationB = ({ text }) => <div style={{ padding: '1rem', background: '#e0f0ff', border: '1px solid #0066cc' }}>InformationB: {text}</div>;
-}
-
-function MyComponentsTest() {
-  return (
-    <div style={{ padding: '2rem', fontFamily: 'Arial, sans-serif' }}>
-      <h1>My Components Test Page</h1>
-      
-      <section style={{ marginBottom: '2rem', padding: '1rem', border: '1px solid #ccc', borderRadius: '4px' }}>
-        <h2>CheckboxGroup Component (Built-in)</h2>
-        <CheckboxGroup />
-        <h3>Vertical Layout</h3>
-        <CheckboxGroup horizontal={false} />
-        <h3>Custom Data</h3>
-        <CheckboxGroup data={["Option 1", "Option 2", "Option 3"]} />
-      </section>
-
-      <section style={{ marginBottom: '2rem', padding: '1rem', border: '1px solid #ccc', borderRadius: '4px' }}>
-        <h2>CheckboxGroupTest Component</h2>
-        <CheckboxGroupTest />
-      </section>
-
-      <section style={{ marginBottom: '2rem', padding: '1rem', border: '1px solid #ccc', borderRadius: '4px' }}>
-        <h2>InformationTest Component</h2>
-        <InformationTest />
-      </section>
-
-      <section style={{ marginBottom: '2rem', padding: '1rem', border: '1px solid #ccc', borderRadius: '4px' }}>
-        <h2>Alert Components</h2>
-        <AlertB text="This is an alert message." />
-        <InformationB text="This is an informational message." />
-      </section>
-    </div>
-  );
-}
-
-console.log('Index.js loading...');
 
 ReactDOM.createRoot(document.getElementById('root')).render(
     <React.StrictMode>
-        <MyComponentsTest />
+        <App />
     </React.StrictMode>
 );
 
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
